requirements/common.txt
Django
djangorestframework
git+https://github.com/Elgismarus/postalcode@v0.0.1#egg=postalcode
